refactor: 기존 LottoApplication, LottoMachine 기능 구현한 클래스로 대체
